refactor of tests with a db reset step added to teardown
